action: Lock rust version to 1.52
Got the following error in the clippy:
error: this expression borrows a reference (`&str`) that is immediately dereferenced by the compiler
  --> tests/cpuset.rs:54:37
   |
54 |             let r = cpuset.set_cpus(&cpus);
   |                                     ^^^^^ help: change this to: `cpus`
   |
   = note: `-D clippy::needless-borrow` implied by `-D warnings`
   = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#needless_borrow

But we did not touch the code, just becuase the clippy rules was updated
follow the rust update.

To avoid this problem we lock the rust version in CI and make it
steady.

Signed-off-by: Tim Zhang <tim@hyper.sh>